+++ src/datetime.rs
@@ -7,7 +7,7 @@ use super::{BitFlags, Error, Register, DEVICE_ADDRESS, DS1307};
 use hal::blocking::i2c::{Write, WriteRead};
 
 /// Hours in either 12-hour (AM/PM) or 24-hour format
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Hours {
     /// AM [1-12]
     AM(u8),
@@ -18,7 +18,7 @@ pub enum Hours {
 }
 
 /// Date and time
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub struct DateTime {
     /// Year [2000-2099]
     pub year: u16,