+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Day of month validity check.
 
 ### Changed
-- `Hours`, `DateTime` and `SQWOUTRateBits` are now `PartialEq`.
+- `Hours`, `DateTime` and `SQWOUTRateBits` are now `Copy` and `PartialEq`.
 
 ## [0.2.0] - 2018-09-07